requirements.txt
flask
tensorflow
pillow
requests
